+++ CHANGES.md
@@ -17,6 +17,7 @@ The release versions are PyPi releases.
   * Remove most of `fake_filesystem_shutil` implementation, relying on the patched `os` module instead ([#194](../../issues/194)).
 
 #### Fixes
+  * Renaming a directory to an existing directory was not possible (Posix only, [#217](../../issues/217)).
   * A non-existing file that has the same as the content of the parent object was seen as existent ([#219](../../issues/219)).
   * Incorrect handling of symlinks in `os.path.size` and other functions ([#210](../../issues/210)).
   * Failed to create directory with symlink as parent ([#215](../../issues/215)).